Move and rename song helper function
