--- src/lib/utils/song.ts
@@ -1,9 +1,0 @@
-export function fixSongText(s: string): string {
-  return s
-    .replaceAll("---", "—")
-    .replaceAll("--", "–")
-    .replaceAll("||:", "𝄆")
-    .replaceAll(":||", "𝄇")
-    .replaceAll("|:", "𝄆")
-    .replaceAll(":|", "𝄇");
-}
